+++ lib/libs/bitcoin_lib/lib/src/protocol/types/payload_length.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import '../../encode.dart';
+import '../../utils/encode.dart';
 
 class PayloadLength {
   final int value;
