+++ lib/libs/bitcoin_lib/lib/src/protocol/types/timestamp.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import '../../encode.dart';
+import '../../utils/encode.dart';
 
 class Timestamp {
   final int unixtime;
