+++ lib/libs/bitcoin_lib/lib/src/protocol/types/variable_length_integer.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import '../../encode.dart';
+import '../../utils/encode.dart';
 
 class VarInt {
   late final int length;
