+++ lib/libs/bitcoin_lib/lib/src/protocol/types/version.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import '../../encode.dart';
+import '../../utils/encode.dart';
 
 enum Version {
   // in Bitcoin Core, this value is 70016
