+++ lib/libs/bitcoin_lib/lib/src/utils/encode.dart
