+++ lib/libs/bitcoin_lib/lib/src/utils/hash.dart
